Maps of regional price increases
Regional maximum percentage increase in prices during a twelve-month period with respect to the initial price due to a 1% increase in price in region coloured dark blue. Values are based on alpha=0, posterior samples of beta_{i,j} and gamma_{i,j,t}, with t fixed to July of the year marked in each panel. The striped gray area is not included in our study.